# Modules/Purchase/routes/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Purchase Management Module API",
    "description": "Comprehensive Purchase Management API for Branch Managers - Handles order creation, tracking, receiving goods, and return management",
    "version": "1.0.0",
    "contact": {
      "name": "Assab API Support"
    }
  },
  "servers": [
    {
      "url": "/api/v1/purchase",
      "description": "Purchase Module API Base URL"
    }
  ],
...
  "paths": {
...
    "/history/{id}": {
      "get": {
        "tags": ["Purchase History"],
        "summary": "Get order details",
        "description": "Retrieve detailed information about a specific order",
        "operationId": "getOrderDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
    "/orders/compare-prices": {
      "post": {
        "tags": ["New Order"],
        "summary": "Compare prices across sources",
        "description": "Compare prices for an item across Direct Supplier, Via PO, and Internal Transfer",
        "operationId": "comparePrices",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparePricesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price comparison retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceComparisonResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/suppliers": {
      "get": {
        "tags": ["New Order"],
        "summary": "Get suppliers for item",
        "description": "Get list of suppliers offering a specific item",
        "operationId": "getSuppliers",
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by supplier status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["online", "offline", "away"]
            }
          },
          {
            "name": "max_delivery_hours",
            "in": "query",
            "description": "Maximum delivery time in hours",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search by supplier name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suppliers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/branches": {
      "get": {
        "tags": ["New Order"],
        "summary": "Get branches with stock",
        "description": "Get list of branches with available stock for internal transfer",
        "operationId": "getBranches",
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "description": "Required quantity",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "min_availability",
            "in": "query",
            "description": "Minimum availability percentage (60-100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 60,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search by branch name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branches retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/direct-supplier": {
      "post": {
        "tags": ["New Order"],
        "summary": "Create direct supplier order",
        "description": "Create a new order directly from a supplier",
        "operationId": "storeDirectSupplier",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectSupplierOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Direct supplier order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/purchasing-officer": {
      "post": {
        "tags": ["New Order"],
        "summary": "Create via purchasing officer order",
        "description": "Create a new order via purchasing officer",
        "operationId": "storePurchasingOfficer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchasingOfficerOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Purchasing officer order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/internal-transfer": {
      "post": {
        "tags": ["New Order"],
        "summary": "Create internal transfer order",
        "description": "Create a new internal transfer order from another branch",
        "operationId": "storeInternalTransfer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalTransferRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Internal transfer order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/orders/draft": {
      "post": {
        "tags": ["New Order"],
        "summary": "Save order as draft",
        "description": "Save an order as draft for later submission",
        "operationId": "saveDraft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order saved as draft successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
    "/orders/{id}/submit": {
      "post": {
        "tags": ["New Order"],
        "summary": "Submit order",
        "description": "Submit a draft order for processing",
        "operationId": "submitOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot submit order in current status"
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
    "/suppliers": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Get suppliers list",
        "description": "Retrieve list of suppliers with filtering options",
        "operationId": "getSuppliers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["online", "offline", "away"]
            }
          },
          {
            "name": "available_only",
            "in": "query",
            "description": "Show only available suppliers",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "max_delivery_hours",
            "in": "query",
            "description": "Maximum delivery time",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "min_rating",
            "in": "query",
            "description": "Minimum rating",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search by supplier name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suppliers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSuppliers"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object",
            "properties": {
              "pagination": {
                "type": "object"
              },
              "filters_applied": {
                "type": "object"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "object"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ComparePricesRequest": {
        "type": "object",
        "required": ["item_id", "quantity"],
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "format": "float",
            "minimum": 0.001
          },
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "PriceComparisonResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "item_id": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "sources": {
                    "type": "object",
                    "properties": {
                      "direct_supplier": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "via_purchasing_officer": {
                        "type": "object",
                        "nullable": true
                      },
                      "internal_transfer": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      }
                    }
                  },
                  "best_option": {
                    "type": "object",
                    "nullable": true
                  },
                  "insights": {
                    "type": "object",
                    "properties": {
                      "lowest_price": {
                        "type": "object",
                        "nullable": true
                      },
                      "fastest_delivery": {
                        "type": "object",
                        "nullable": true
                      },
                      "best_rating": {
                        "type": "object",
                        "nullable": true
                      },
                      "best_compliance": {
                        "type": "object",
                        "nullable": true
                      }
                    }
                  },
                  "price_trends": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "DirectSupplierOrderRequest": {
        "type": "object",
        "required": ["supplier_id", "quality_level", "notification_channels", "items"],
        "properties": {
          "supplier_id": {
            "type": "string",
            "format": "uuid"
          },
          "quality_level": {
            "type": "string",
            "enum": ["economy", "standard", "premium"]
          },
          "notification_channels": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": ["email", "whatsapp", "app", "sms"]
            }
          },
          "message": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["item_name", "quantity", "unit", "unit_price"],
              "properties": {
                "item_id": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "item_name": {
                  "type": "string",
                  "maxLength": 255
                },
                "item_logo": {
                  "type": "string",
                  "nullable": true
                },
                "quantity": {
                  "type": "number",
                  "format": "float",
                  "minimum": 0.001
                },
                "unit": {
                  "type": "string",
                  "enum": ["kg", "pk", "unit", "box", "liter", "piece"]
                },
                "unit_price": {
                  "type": "number",
                  "format": "float",
                  "minimum": 0
                },
                "quality": {
                  "type": "string",
                  "enum": ["economy", "standard", "premium"],
                  "nullable": true
                }
              }
            }
          }
        }
      },
      "PurchasingOfficerOrderRequest": {
        "type": "object",
        "required": ["quality_level", "processing_time", "preferred_delivery_date", "latest_delivery_date", "items"],
        "properties": {
          "quality_level": {
            "type": "string",
            "enum": ["economy", "standard", "premium"]
          },
          "processing_time": {
            "type": "string",
            "enum": ["standard", "urgent"]
          },
          "preferred_delivery_date": {
            "type": "string",
            "format": "date"
          },
          "latest_delivery_date": {
            "type": "string",
            "format": "date"
          },
          "special_instructions": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "message": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "InternalTransferRequest": {
        "type": "object",
        "required": ["from_branch_id", "priority", "items"],
        "properties": {
          "from_branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority": {
            "type": "string",
            "enum": ["high", "normal"]
          },
          "message": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["item_name", "quantity", "unit"],
              "properties": {
                "item_id": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "item_name": {
                  "type": "string",
                  "maxLength": 255
                },
                "quantity": {
                  "type": "number",
                  "format": "float",
                  "minimum": 0.001
                },
                "unit": {
                  "type": "string",
                  "enum": ["kg", "pk", "unit", "box", "liter", "piece"]
                },
                "quality": {
                  "type": "string",
                  "enum": ["economy", "standard", "premium"],
                  "nullable": true
                },
                "available_in_source": {
                  "type": "number",
                  "format": "float",
                  "nullable": true
                },
                "expiry_date": {
                  "type": "string",
                  "format": "date",
                  "nullable": true
                },
                "cooling_status": {
                  "type": "boolean",
                  "nullable": true
                }
              }
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "required": ["item_name", "quantity", "unit", "unit_price"],
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "item_name": {
            "type": "string",
            "maxLength": 255
          },
          "item_logo": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "float",
            "minimum": 0.001
          },
          "unit": {
            "type": "string",
            "enum": ["kg", "pk", "unit", "box", "liter", "piece"]
          },
          "unit_price": {
            "type": "number",
            "format": "float",
            "minimum": 0
          },
          "quality": {
            "type": "string",
            "enum": ["economy", "standard", "premium"],
            "nullable": true
          }
        }
      },
      "OrderDraftRequest": {
        "type": "object",
        "properties": {
          "order_type": {
            "type": "string",
            "enum": ["direct_supplier", "via_purchasing_officer", "internal_transfer"]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
...
      "PaginatedSuppliers": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          }
        ]
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["online", "offline", "away"]
          },
          "status_label": {
            "type": "string",
            "example": "Online"
          },
          "is_available": {
            "type": "boolean"
          },
          "contact_methods": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["email", "whatsapp", "app", "sms"]
            }
          },
          "default_delivery_hours": {
            "type": "integer",
            "nullable": true
          },
          "average_response_time_hours": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "response_rate_percentage": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      }
    }
  }
}